refactor: move getBranchAndIndices to lib file
